requirements.txt
ephem==3.7.7.0
python-telegram-bot==12.2.0
PySocks==1.7.1